test(time-range): move tests relevant to x-axis are uPlot plugin only
